Revert "Restore avatar after user logs out from sync."
This reverts commit ec686dc0ec5b8706a01dad1663f80f771da8a269.

Reason for revert:
The fix does not work in all cases and it adds a lot of complexity in
the way avatar images are now used. There are now 3 sources of truth for
the avatar icon index and it is not clear what the relationship between
them is:
* kAccountIdKey in the profile entry cache
* kProfileAvatarIndex which is a synced pref.
* kLocalProfileAvatarIndex which is non-syncable pref.

Note that kProfileAvatarIndex and kLocalProfileAvatarIndex are both set
when the user updates the image icon in chrome://settings/manageProfile.

I've just tested and if I start sync, set a local avatar icon, then turn
off sync, I still have the same image. So this CL does not fix the bug
863008 for this scenario.

The core issue comes from the fact that we try to fix a problem that
cannot be fixed in the current architecture of sync - sync preferences
do not have a local variant and do not get reverted to a local variant when
the user turns sync off. I think doing a custom implementation for the profile
avatar (and not for the profile name for example or settings) just adds confusion.

The fact that the relationship between these preferences is not very clear,
fixing the issue 900374 becomes hard and is risky.

Bug: 900374

Original change's description:
> Restore avatar after user logs out from sync.
>
> BUG=863008
>
> Change-Id: If083bc0703e251fb6e9c597fe5b7f842e37875c9
>
> R=​bauerb@chromium.org,sky@chromium.org
>
> Change-Id: If083bc0703e251fb6e9c597fe5b7f842e37875c9
> Reviewed-on: https://chromium-review.googlesource.com/1135137
> Commit-Queue: Thomas Tangl <tangltom@chromium.org>
> Reviewed-by: Bernhard Bauer <bauerb@chromium.org>
> Reviewed-by: Thomas Tangl <tangltom@chromium.org>
> Reviewed-by: David Roger <droger@chromium.org>
> Reviewed-by: Scott Violet <sky@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#582541}

TBR=sky@chromium.org,droger@chromium.org,bauerb@chromium.org,tangltom@chromium.org,boocmp@yandex-team.ru

(cherry picked from commit 8b49dd022282696d499bd6cdf2ebee7a2547f5cd)

Bug: 863008
Change-Id: I46aed023494d94e197828339f6870bf9ba0c9da8
Reviewed-on: https://chromium-review.googlesource.com/c/1329153
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Thomas Tangl <tangltom@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: anthonyvd <anthonyvd@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#607176}
Reviewed-on: https://chromium-review.googlesource.com/c/1338102
Cr-Commit-Position: refs/branch-heads/3578@{#691}
Cr-Branched-From: 4226ddf99103e493d7afb23a4c7902ee496108b6-refs/heads/master@{#599034}